+++ packages/cubejs-schema-compiler/src/adapter/BaseDimension.ts
@@ -1,28 +1,40 @@
+import type { BaseQuery } from './BaseQuery';
+import type { DimensionDefinition, SegmentDefinition } from '../compiler/CubeEvaluator';
+
 export class BaseDimension {
-  constructor(query, dimension) {
-    this.query = query;
+  public readonly expression: any;
+
+  public readonly expressionCubeName: any;
+
+  public readonly expressionName: string | undefined;
+
+  public readonly isMemberExpression: boolean = false;
+
+  public constructor(
+    protected readonly query: BaseQuery,
+    public readonly dimension: any
+  ) {
     if (dimension && dimension.expression) {
       this.expression = dimension.expression;
       this.expressionCubeName = dimension.cubeName;
       this.expressionName = dimension.expressionName || `${dimension.cubeName}.${dimension.name}`;
       this.isMemberExpression = !!dimension.definition;
     }
-    this.dimension = dimension;
   }
 
-  selectColumns() {
+  public selectColumns(): string[] | null {
     return [`${this.dimensionSql()} ${this.aliasName()}`];
   }
 
-  hasNoRemapping() {
+  public hasNoRemapping() {
     return this.dimensionSql() === this.aliasName();
   }
 
-  cumulativeSelectColumns() {
+  public cumulativeSelectColumns() {
     return [`${this.aliasName()}`];
   }
 
-  dimensionSql() {
+  public dimensionSql() {
     if (this.expression) {
       return this.convertTzForRawTimeDimensionIfNeeded(() => this.query.evaluateSymbolSql(this.expressionCubeName, this.expressionName, this.definition(), 'dimension'));
     }
@@ -32,7 +44,7 @@ export class BaseDimension {
     return this.convertTzForRawTimeDimensionIfNeeded(() => this.query.dimensionSql(this));
   }
 
-  convertTzForRawTimeDimensionIfNeeded(sql) {
+  public convertTzForRawTimeDimensionIfNeeded(sql) {
     if (this.query.options.convertTzForRawTimeDimension) {
       return this.query.evaluateSymbolSqlWithContext(sql, {
         convertTzForRawTimeDimension: true
@@ -42,66 +54,71 @@ export class BaseDimension {
     }
   }
 
-  sqlDefinition() {
+  public sqlDefinition() {
     return this.dimensionDefinition().sql;
   }
 
-  getMembers() {
+  public getMembers() {
     return [this];
   }
 
-  cube() {
+  public cube() {
     if (this.expression) {
       return this.query.cubeEvaluator.cubeFromPath(this.expressionCubeName);
     }
     return this.query.cubeEvaluator.cubeFromPath(this.dimension);
   }
 
-  dimensionDefinition() {
+  public dimensionDefinition(): DimensionDefinition | SegmentDefinition {
     if (this.query.cubeEvaluator.isSegment(this.dimension)) {
       return this.query.cubeEvaluator.segmentByPath(this.dimension);
     }
+
     return this.query.cubeEvaluator.dimensionByPath(this.dimension);
   }
 
-  definition() {
+  public definition(): DimensionDefinition | SegmentDefinition {
     if (this.expression) {
       return {
         sql: this.expression,
         // TODO use actual dimension type even though it isn't used right now
         type: 'number'
-      };
+      } as DimensionDefinition;
     }
+
     return this.dimensionDefinition();
   }
 
-  aliasName() {
+  public aliasName(): string | null {
     // Require should be here because of cycle depend
     return this.query.escapeColumnName(this.unescapedAliasName());
   }
 
-  unescapedAliasName() {
-    if (this.expression) {
+  public unescapedAliasName(): string {
+    if (this.expression && this.expressionName) {
       return this.query.aliasName(this.expressionName);
     }
+
     return this.query.aliasName(this.dimension);
   }
 
-  dateFieldType() {
+  public dateFieldType() {
     return this.dimensionDefinition().fieldType;
   }
 
-  path() {
+  public path() {
     if (this.expression) {
       return null;
     }
+
     if (this.query.cubeEvaluator.isSegment(this.dimension)) {
       return this.query.cubeEvaluator.parsePath('segments', this.dimension);
     }
+
     return this.query.cubeEvaluator.parsePath('dimensions', this.dimension);
   }
 
-  expressionPath() {
+  public expressionPath() {
     if (this.expression) {
       return `expr:${this.expression.expressionName}`;
     }